Script tunning
