+++ scripts/requirements.txt
@@ -1,0 +1,10 @@
+biplist==1.0.3
+colorama==0.4.4
+cryptography==37.0.2
+dicttoxml==1.7.4
+iOSbackup==0.9.923
+PyJWT==2.4.0
+requests==2.27.1
+tabulate==0.8.9
+termcolor==1.1.0
+urllib3==1.26.9
